feat: core javascript script refactor in typescript
