+++ package.json
@@ -24,6 +24,7 @@
     "node-fetch": "^2.6.7"
   },
   "devDependencies": {
+    "@types/node-fetch": "^2.6.11",
     "eslint": "^8.55.0",
     "eslint-config-google": "^0.14.0",
     "jest": "^29.7.0",
