+++ src/index.ts
@@ -1,0 +1,1 @@
+export { default as MistralClient } from './mistral-client';
