+++ src/utils/api-error.ts
@@ -1,0 +1,6 @@
+export class MistralAPIError extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = 'MistralAPIError';
+  }
+}
